test(ui): align mobile create coverage
